Refactor copy pasted code for tests
